--- src/main/java/soma/edupi/user/domain/Role.java
@@ -1,5 +1,0 @@
-package soma.edupi.user.domain;
-
-public enum Role {
-    ROLE_USER
-}
