--- src/main/java/soma/edupi/user/dto/request/EmailRequest.java
@@ -1,16 +1,0 @@
-package soma.edupi.user.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class EmailRequest {
-
-    private String email;
-
-    public EmailRequest(String email) {
-        this.email = email;
-    }
-
-}
