--- src/main/java/soma/edupi/user/exception/DbValidException.java
@@ -1,8 +1,0 @@
-package soma.edupi.user.exception;
-
-public class DbValidException extends RuntimeException {
-
-    public DbValidException(String message) {
-        super(message);
-    }
-}
