--- src/main/java/soma/edupi/user/exception/InnerServerException.java
@@ -1,9 +1,0 @@
-package soma.edupi.user.exception;
-
-public class InnerServerException extends RuntimeException {
-
-    public InnerServerException(String message) {
-        super(message);
-    }
-
-}
