+++ src/main/java/soma/edupiuser/EdupiUserApplication.java
@@ -1,0 +1,15 @@
+package soma.edupiuser;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@EnableAsync
+@SpringBootApplication
+public class EdupiUserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EdupiUserApplication.class, args);
+    }
+
+}
