+++ src/main/java/soma/edupiuser/account/models/AccountLoginRequest.java
@@ -1,0 +1,26 @@
+package soma.edupiuser.account.models;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class AccountLoginRequest {
+
+    @NotNull
+    @NotBlank
+    private String email;
+
+    @NotNull
+    @NotBlank
+    private String password;
+
+    @Builder
+    public AccountLoginRequest(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+}
