+++ src/main/java/soma/edupiuser/account/models/SignupResponse.java
@@ -1,0 +1,15 @@
+package soma.edupiuser.account.models;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class SignupResponse {
+
+    String message;
+
+    public SignupResponse(String message) {
+        this.message = message;
+    }
+}
