+++ src/main/java/soma/edupiuser/account/service/domain/Role.java
@@ -1,0 +1,5 @@
+package soma.edupiuser.account.service.domain;
+
+public enum Role {
+    ROLE_USER
+}
