+++ src/main/java/soma/edupiuser/web/exception/DbValidException.java
@@ -1,0 +1,8 @@
+package soma.edupiuser.web.exception;
+
+public class DbValidException extends RuntimeException {
+
+    public DbValidException(String message) {
+        super(message);
+    }
+}
