+++ src/main/java/soma/edupiuser/web/exception/InnerServerException.java
@@ -1,0 +1,9 @@
+package soma.edupiuser.web.exception;
+
+public class InnerServerException extends RuntimeException {
+
+    public InnerServerException(String message) {
+        super(message);
+    }
+
+}
