+++ src/main/java/soma/edupiuser/web/models/ErrorResponse.java
@@ -1,0 +1,15 @@
+package soma.edupiuser.web.models;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ErrorResponse {
+
+    String message;
+
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+}
