+++ src/test/java/soma/edupiuser/EdupiUserApplicationTests.java
@@ -1,0 +1,13 @@
+package soma.edupiuser;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EdupiUserApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
